Crate iota_build_cache_server

IOTA Build Cache Server Library

This crate provides both a standalone build cache server binary and a client library for interacting with build cache servers.

Re-exports

pub use client::BuildCacheClient;

pub use client::BuildCacheError;

pub use types::BuildCacheResponse;

pub use types::BuildJob;

pub use types::BuildRequest;

pub use types::BuildStatus;

Modules

cache

client

server

types